Fixed timeout_handler, added documentation, fixed pipe param

+++ src/autograder.c
@@ -13,20 +13,24 @@ int total_params;         // Total number of parameters to test - (argc - 2)
 // Contains status of child processes (-1 for done, 1 for still running)
 int *child_status;
 
-    /* TODO: 
-        * Begin Change 4 in mq_autograder.c
-        * Implement get_score()
-    */
-
 // TODO (Change 3): Timeout handler for alarm signal - kill remaining running child processes
 void timeout_handler(int signum) {
-    pid_t pid;
-    int status;
+    // Kill all child processes that are still running
 
-    // Kill everything 
+    /*
+    If you are failing the timeout_handler() test, make sure you are checking 
+    if child_status[i] == 1 before calling kill() on pid[i]. If our tests, 
+    pid[i] = -1 if child_status[i] == -1, so calling kill(pid[i], ...) 
+    will result in errors since kill() will kill all processes available if 
+    the first argument is -1.
+    */
+
     for (int i = 0; i < curr_batch_size; ++i) {
-        kill(pids[i], SIGKILL);
-    }
+        if (child_status[i] == 1 && pids[i] > 0) {
+            kill(pids[i], SIGKILL);   
+            child_status[i] = -1; 
+        }
+    } 
 }
 
 // Execute the student's executable using exec()
@@ -44,15 +48,12 @@ void execute_solution(char *executable_path, char *input, int batch_idx) {
 
     // Child process
     if (pid == 0) {
-        char *executable_name = get_exe_name(executable_path);
-
-        int print_fd = STDOUT_FILENO;
-
+        char *executable_name = get_exe_name(executable_path); // Get the executable name without the path
         // TODO (Change 1): Redirect STDOUT to output/<executable>.<input> file
         char output_file[BUFSIZ];
-        sprintf(output_file, "output/%s.%s", executable_name, input);
-        int output_fd = open(output_file, O_WRONLY | O_CREAT | O_TRUNC, 0666);
-        if (output_fd == -1) {  
+        sprintf(output_file, "output/%s.%s", executable_name, input); // Create the output file path
+        int output_fd = open(output_file, O_WRONLY | O_CREAT | O_TRUNC, 0666); // Open the output file, create if it doesn't exist
+        if (output_fd == -1) {
             perror("open");
             exit(1);
         }
@@ -66,42 +67,36 @@ void execute_solution(char *executable_path, char *input, int batch_idx) {
         // TODO (Change 2): Handle different cases for input source
         #ifdef EXEC
                     
-        // printf("%s %s %s\n", executable_path, executable_name, input);
-        execlp(executable_path, executable_name, input, (char *) NULL);
+        execlp(executable_path, executable_name, input, (char *) NULL); // Execute the program with the input parameter
         
         #elif REDIR
             
         // TODO: Redirect STDIN to input/<input>.in file
 
         char input_file[BUFSIZ];
-        sprintf(input_file, "input/%s.in", input);
-        printf("%s\n", input_file);
-        int input_fd = open(input_file, O_RDONLY | O_CREAT, 0666);
+        sprintf(input_file, "input/%s.in", input); // Create the input file path
+        int input_fd = open(input_file, O_RDONLY | O_CREAT, 0666); // Open the input file, create if it doesn't exist
 
         if (input_fd == -1) {  
             perror("open");
             exit(1);
         }
 
-        if (dup2(input_fd, STDIN_FILENO) == -1) {
+        if (dup2(input_fd, STDIN_FILENO) == -1) { // Redirect STDIN to the input file
             perror("dup2");
             close(input_fd);
             exit(1);
         }
         
-        // printf("%s %s %s\n", executable_path, executable_name, input);
-        execlp(executable_path, executable_name, (char *) NULL);
+        execlp(executable_path, executable_name, (char *) NULL); // Execute the program without the input parameter
 
         #elif PIPE
             // TODO: Pass read end of pipe to child process 
         close(pipefd[1]); // close write end of pipe
-        // write(STDERR_FILENO, &pipefd[0], sizeof(pipefd[0]));
-
-        // printf("%s\n", pipefd[0]);
         char fd_str[10];  // Buffer to hold the string representation of the file descriptor
         sprintf(fd_str, "%d", pipefd[0]);  // Convert the file descriptor to a string
         
-        execlp(executable_path, executable_name, fd_str, (char *) NULL);
+        execlp(executable_path, executable_name, fd_str, (char *) NULL); // Execute the program with the pipe file descriptor
         
         #endif
 
@@ -154,8 +149,6 @@ void monitor_and_evaluate_solutions(int tested, char *param, int param_idx) {
         }
 
         // TODO: Determine if the child process finished normally, segfaulted, or timed out
-        int exit_status = WEXITSTATUS(status);
-        int exited = WIFEXITED(status);
         int signaled = WIFSIGNALED(status);
 
         // Programs either exit with a status or are killed by a signal
@@ -173,19 +166,15 @@ void monitor_and_evaluate_solutions(int tested, char *param, int param_idx) {
 
         // TODO: Also, update the results struct with the status of the child process
         char output_file[BUFSIZ];
-        char* filename = get_exe_name(results[tested - curr_batch_size + j].exe_path);
-        // strrchr(results[tested - curr_batch_size + j].exe_path, '/');
-        // if (filename != NULL) {
-        //     ++filename;
-        // }
-        sprintf(output_file, "output/%s.%s", filename, param);
-        int output_fd = open(output_file, O_RDONLY);
+        char* filename = get_exe_name(results[tested - curr_batch_size + j].exe_path); // Get the executable name without the path
+        sprintf(output_file, "output/%s.%s", filename, param); // Create the output file path
+        int output_fd = open(output_file, O_RDONLY); // Open the output file
         if (output_fd == -1) {  
             perror("open");
             exit(1);
         }
         char buffer[BUFSIZ];
-        ssize_t num_bytes = read(output_fd, buffer, sizeof(buffer));
+        ssize_t num_bytes = read(output_fd, buffer, sizeof(buffer)); // Read the output file
         if (num_bytes == -1) {
             perror("read");
             exit(1);
@@ -193,13 +182,12 @@ void monitor_and_evaluate_solutions(int tested, char *param, int param_idx) {
 
         if (num_bytes != 0) {
             buffer[num_bytes] = '\0';  // Null-terminate the buffer
-            results[tested - curr_batch_size + j].status[param_idx] = atoi(buffer);
+            results[tested - curr_batch_size + j].status[param_idx] = atoi(buffer) + 1; // Convert from 0/1 to 1/2 to match enum
         }
         close(output_fd);
 
         // NOTE: Make sure you are using the output/<executable>.<input> file to determine the status
         //       of the child process, NOT the exit status like in Project 1.
-
 
         // Adding tested parameter to results struct
         results[tested - curr_batch_size + j].params_tested[param_idx] = atoi(param);
@@ -284,7 +272,7 @@ int main(int argc, char *argv[]) {
     write_results_to_file(results, num_executables, total_params);
 
     // You can use this to debug your scores function
-    get_score("results.txt", results[0].exe_path);
+    // get_score("results.txt", results[0].exe_path);
 
     // Print each score to scores.txt
     write_scores_to_file(results, num_executables, "results.txt");
