+++ src/worker.c
@@ -25,9 +25,15 @@ long worker_id;        // Used for sending/receiving messages from the message q
 void timeout_handler(int signum) {
     // Kill everything 
     // write(STDERR_FILENO, "timed out\n", 10);
+    // for (int i = 0; i < curr_batch_size; ++i) {
+    //     kill(pids[i], SIGKILL);
+    // }
     for (int i = 0; i < curr_batch_size; ++i) {
-        kill(pids[i], SIGKILL);
-    }
+        if (child_status[i] == 1 && pids[i] > 0) {
+            kill(pids[i], SIGKILL);   
+            child_status[i] = -1; 
+        }
+    } 
 }
 
 
@@ -60,10 +66,10 @@ void execute_solution(char *executable_path, int param, int batch_idx) {
         memset(param_str, 0, sizeof(param_str));
         sprintf(param_str, "%d", param);
 
-        char exe_sol_msg[BUFSIZ];
-        memset(exe_sol_msg, 0, BUFSIZ);
-        sprintf(exe_sol_msg, "executing: %s, param: %d, param_str: %s\n", executable_name, param, param_str);
-        write(STDERR_FILENO, exe_sol_msg, strlen(exe_sol_msg));
+        // char exe_sol_msg[BUFSIZ];
+        // memset(exe_sol_msg, 0, BUFSIZ);
+        // sprintf(exe_sol_msg, "executing: %s, param: %d, param_str: %s\n", executable_name, param, param_str);
+        // write(STDERR_FILENO, exe_sol_msg, strlen(exe_sol_msg));
 
         execlp(executable_path, executable_name, param_str, (char *) NULL);
 
@@ -160,7 +166,7 @@ void monitor_and_evaluate_solutions(int finished) {
 
         if (num_bytes != 0) {
             buffer[num_bytes] = '\0';  // Null-terminate the buffer
-            pairs[finished + j].status = atoi(buffer);
+            pairs[finished + j].status = atoi(buffer) + 1; // Convert from 0/1 to 1/2 to match enum
         }
         close(output_fd);
 
@@ -222,16 +228,16 @@ int main(int argc, char **argv) {
         exit(1);
     }
 
-    write(STDERR_FILENO, "worker received msg\n", 20);
+    // write(STDERR_FILENO, "worker received msg\n", 20);
 
     // TODO: Parse message and set up pairs_t array
     int pairs_to_test = atoi(msg.mtext);
     pairs = malloc(sizeof(pairs_t) * pairs_to_test);
 
-    char pairs_str[128];
-    memset(pairs_str, 0, 128);
-    sprintf(pairs_str, "pairs to test: %d\n", pairs_to_test);
-    write(STDERR_FILENO, pairs_str, strlen(pairs_str));
+    // char pairs_str[128];
+    // memset(pairs_str, 0, 128);
+    // sprintf(pairs_str, "pairs to test: %d\n", pairs_to_test);
+    // write(STDERR_FILENO, pairs_str, strlen(pairs_str));
 
     // TODO: Receive (executable, parameter) pairs from autograder and store them in pairs_t array.
     //       Messages will have the format ("%s %d", executable_path, parameter). (mtype = worker_id)
@@ -246,7 +252,7 @@ int main(int argc, char **argv) {
         memset(&exec_path, 0, sizeof(exec_path));
 
         if (sscanf(exec_msg.mtext, "%s %d", exec_path, &pairs[i].parameter) != 2) {
-            write(STDERR_FILENO, "error\n", 6);
+            write(STDERR_FILENO, "error receiving pair message\n", 30);
             exit(1);
         }
 
@@ -254,13 +260,13 @@ int main(int argc, char **argv) {
         strcpy(pairs[i].executable_path, exec_path);
         pairs[i].status = 0;
 
-        char msg[BUFSIZ];
-        memset(msg, 0, BUFSIZ);
-        sprintf(msg, "worker %ld received: %s -> exe_path: %s, param: %d\n", worker_id, exec_msg.mtext, pairs[i].executable_path, pairs[i].parameter);
-        write(STDERR_FILENO, msg, strlen(msg));
+        // char msg[BUFSIZ];
+        // memset(msg, 0, BUFSIZ);
+        // sprintf(msg, "worker %ld received: %s -> exe_path: %s, param: %d\n", worker_id, exec_msg.mtext, pairs[i].executable_path, pairs[i].parameter);
+        // write(STDERR_FILENO, msg, strlen(msg));
     }
 
-    write(STDERR_FILENO, "worker all received pairs\n", 27);
+    // write(STDERR_FILENO, "worker all received pairs\n", 27);
 
     // TODO: Send ACK message to mq_autograder after all pairs received (mtype = BROADCAST_MTYPE)
 
@@ -274,7 +280,7 @@ int main(int argc, char **argv) {
         ack.mtype = BROADCAST_MTYPE;
         sprintf(ack.mtext, "%s", "ACK");
 
-        write(STDERR_FILENO, "worker sent ack\n", 16);
+        // write(STDERR_FILENO, "worker sent ack\n", 16);
         if (msgsnd(msqid, &ack, sizeof(msgbuf_t) - sizeof(long), 0) == -1) {
             perror("failed to send acknowledgement"); 
             exit(1);
@@ -290,11 +296,11 @@ int main(int argc, char **argv) {
         if (strcmp(synacks.mtext, "SYNACK") == 0) {
             break;
         }
-        write(STDERR_FILENO, "worker did not receive synack, retrying\n", 41);
+        // write(STDERR_FILENO, "worker did not receive synack, retrying\n", 41);
     }
 
-    write(STDERR_FILENO, "worker received synack\n", 23);
-    write(STDERR_FILENO, "worker getting to work\n", 23);
+    // write(STDERR_FILENO, "worker received synack\n", 23);
+    // write(STDERR_FILENO, "worker getting to work\n", 23);
 
     // Run the pairs in batches of 8 and send results back to autograder
     for (int i = 0; i < pairs_to_test; i+= PAIRS_BATCH_SIZE) {
@@ -312,14 +318,14 @@ int main(int argc, char **argv) {
 
         // TODO: Wait for the batch to finish and check results
         monitor_and_evaluate_solutions(i); // stuck in here
-        write(STDERR_FILENO, "sending results to autograder\n", 30);
+        // write(STDERR_FILENO, "sending results to autograder\n", 30);
 
         // TODO: Cancel the timer if all child processes have finished
         if (child_status == NULL) {
-            char c_msg[128];
-            memset(c_msg, 0, 128);
-            sprintf(c_msg, "canceled timer for worker %ld\n", worker_id);
-            write(STDERR_FILENO, c_msg, strlen(c_msg));
+            // char c_msg[128];
+            // memset(c_msg, 0, 128);
+            // sprintf(c_msg, "canceled timer for worker %ld\n", worker_id);
+            // write(STDERR_FILENO, c_msg, strlen(c_msg));
 
             cancel_timer();
         } 
@@ -333,10 +339,10 @@ int main(int argc, char **argv) {
     // TODO: Send DONE message to autograder to indicate that the worker has finished testing
     send_done_msg(msqid, worker_id);
 
-    char done_msg[128];
-    memset(done_msg, 0, 128);
-    sprintf(done_msg, "worker %ld sent done msg\n", worker_id);
-    write(STDERR_FILENO, done_msg, strlen(done_msg));
+    // char done_msg[128];
+    // memset(done_msg, 0, 128);
+    // sprintf(done_msg, "worker %ld sent done msg\n", worker_id);
+    // write(STDERR_FILENO, done_msg, strlen(done_msg));
 
     // Free the pairs_t array
     for (int i = 0; i < pairs_to_test; i++) {
@@ -344,8 +350,8 @@ int main(int argc, char **argv) {
     }
     free(pairs);
 
-    char ret_msg[128];
-    memset(ret_msg, 0, 128);
-    sprintf(ret_msg, "worker %ld finished; terminating\n", worker_id);
-    write(STDERR_FILENO, ret_msg, strlen(ret_msg));
+    // char ret_msg[128];
+    // memset(ret_msg, 0, 128);
+    // sprintf(ret_msg, "worker %ld finished; terminating\n", worker_id);
+    // write(STDERR_FILENO, ret_msg, strlen(ret_msg));
 }
